% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodNA.R
\name{prodNA}
\alias{prodNA}
\title{Introduce Missing Values Completely at Random}
\usage{
prodNA(x, noNA = 0.1)
}
\arguments{
\item{x}{dataframe subjected to missing value introduction.}

\item{noNA}{proportion of missing values w.r.t. the number of entries of 'x'.}
}
\value{
dataframe with missing values.
}
\description{
'prodNA' artificially introduces missing values. Entries in the given
dataframe are deleted completely at random up to the specified amount.
}
\note{
This function is internally used by \code{\link{missForest}} and
\code{\link{mixError}}.
}
\examples{
\donttest{
data(iris)
## Introduce 5\% of missing values to the iris data set
iris.mis <- prodNA(iris, 0.05)
summary(iris.mis)
}
}
\seealso{
\code{\link{missForest}}
}
\author{
Daniel J. Stekhoven
}
\keyword{NA}
\keyword{classes,}
