% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyStats.R
\name{TidyErrAll}
\alias{TidyErrAll}
\title{Tidy error from true data during iterations}
\usage{
TidyErrAll(obj)
}
\arguments{
\item{obj}{Input object from `missForestRanger` or `missForestSrc`.}
}
\value{
A data frame containing error from true data.
}
\description{
Tidy error from true data during iterations
}
