% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyStats.R
\name{TidyOobErrAll}
\alias{TidyOobErrAll}
\title{Tidy out-of-bag error estimates during iterations}
\usage{
TidyOobErrAll(obj)
}
\arguments{
\item{obj}{Input object from `missForestRanger` or `missForestSrc`.}
}
\value{
A data frame containing out-of-bag error estimates during iterations.
}
\description{
Tidy out-of-bag error estimates during iterations
}
