% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varClass.R
\name{varClass}
\alias{varClass}
\title{Extract Variable Types from a Dataframe}
\usage{
varClass(x)
}
\arguments{
\item{x}{data frame with variables in the columns.}
}
\value{
a vector of length p where p denotes the number of columns in 'x'. The
entries are "numeric" for continuous variables and "factor" for
categorical variables.
}
\description{
'varClass' returns the variable types of a dataframe. It is used internally
in several functions of the 'missForest'-package.
}
\note{
This function is internally used by \code{\link{missForest}} and
\code{\link{mixError}}.
}
\examples{
\donttest{
data(iris)
varClass(iris)
## We have four continuous and one categorical variable.
}
}
\references{
Oba et al. (2003), 'A Bayesian missing value estimation method for gene
expression profile data', Bioinformatics, 19(16), 2088-2096
}
\seealso{
\code{\link{missForest}}, \code{\link{mixError}}, \code{\link{nrmse}}
}
\author{
Daniel J. Stekhoven
}
\keyword{classes}
